+++ assets/components/backupmodx/connector.php
@@ -8,7 +8,7 @@
  * @var modX $modx
  */
 
-require_once dirname(dirname(dirname(dirname(__FILE__)))) . '/config.core.php';
+require_once dirname(__FILE__, 4) . '/config.core.php';
 require_once MODX_CORE_PATH . 'config/' . MODX_CONFIG_KEY . '.inc.php';
 require_once MODX_CONNECTORS_PATH . 'index.php';