+++ core/components/backupmodx/model/backupmodx/backupmodx.class.php
@@ -2,12 +2,12 @@
 /**
  * BackupMODX
  *
- * Copyright 2015-2023 by Jan Dähne <info@quadro-system.de>
+ * Copyright 2015-2024 by Jan Dähne <info@quadro-system.de>
  *
  * @package backupmodx
  */
 
-require_once dirname(dirname(__DIR__)) . '/vendor/autoload.php';
+require_once dirname(__DIR__, 2) . '/vendor/autoload.php';
 
 /**
  * Class BackupMODX