+++ core/components/backupmodx/src/Helper/SQLImport.php
@@ -2,7 +2,7 @@
 /**
  * SQLImport
  *
- * This script imports SQL commands "as is". Double check the SQL commands
+ * This script imports SQL commands "as is". Double-check the SQL commands
  * before using it and BACKUP YOUR DATABASE.
  *
  * Based on https://gist.github.com/b4oshany/0698d9f32589b77abdcb
@@ -24,7 +24,7 @@ class SQLImport
      * @param modX $modx A reference to the modX instance.
      * @param string $sqlFile The filename of the file containing the mysql-dump data.
      * @param bool $logQuery Return a log of all executed queries.
-     * @return bool|string|array Returns true or ["success" => true, "log" => $log], if the SQL was imported successfully.
+     * @return string|true Returns true or ["success" => true, "log" => $log], if the SQL was imported successfully.
      * @throws Exception
      */
     public static function importSQL(modX $modx, $sqlFile, $logQuery = false)
@@ -33,11 +33,11 @@ class SQLImport
         $file = file($sqlFile);
 
         # import file line by line
-        # and filter (remove) those lines, beginning with an sql comment token
+        # and filter (remove) those lines, beginning with a sql comment token
         $file = array_filter($file, function ($line) {
             return strpos(ltrim($line), '--') !== 0;
         });
-        # and filter (remove) those lines, beginning with an sql notes token
+        # and filter (remove) those lines, beginning with a sql notes token
         $file = array_filter($file, function ($line) {
             return strpos(ltrim($line), '/*') !== 0;
         });
@@ -60,7 +60,7 @@ class SQLImport
                         $sql .= $query . ' ';
                         $del_num = false;
                         if ($logQuery) {
-                            $log .= "{$sql}\ndo---do\n";
+                            $log .= "$sql\ndo---do\n";
                         }
                         $stmt = $modx->prepare($sql);
                         if (!$stmt->execute()) {
@@ -74,12 +74,12 @@ class SQLImport
             } else {
                 $delimiter = is_int(strpos($query, ';'));
                 if ($delimiter) {
-                    $stmt = $modx->prepare("{$sql} {$query}");
+                    $stmt = $modx->prepare("$sql $query");
                     if (!$stmt->execute()) {
                         throw new Exception(implode(' ', $stmt->errorInfo()) . "\n" . 'SQL Query: ' . $query);
                     }
                     if ($logQuery) {
-                        $log .= "{$sql} {$query}\n---\n";
+                        $log .= "$sql $query\n---\n";
                     }
                     $sql = '';
                 } else {