+++ core/components/backupmodx/docs/changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.1] - 2024-04-16
+
+### Added
+
+- Log the size of each file in the backup too
+
+### Fixed
+
+- Trim a trailing slash from excluded folders and exclude those folders too
+
 ## [3.1.0] - 2023-01-13
 
 ### Added
@@ -60,7 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Complete rewrite
 - Use ifsnop/mysqldump-php and Chumper/Zipper composer packages
-- Debug option logging the files to backup and other information in backupmodx.log
+- Debug option logging the files to back up and other information in backupmodx.log
 - Completely AJAX driven dashboard widget
 - Every text displayed in the browser is translateable
 - Catch an error during php shutdown to display (i.e. maximum execution time) errors